fix: enable default system umask for workspace
I have added a `post-start.sh` script to remove ACLs on files in the
workspace mount so that the default system umask is respected.

This change is important because the primary purpose of this project is
to copy files to the user's home directory, and the ownership and
permissions should match the rest of the system.